requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
astroid==3.2.2
attrs==23.2.0
autopep8==2.0.4
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dill==0.3.8
Django==5.0.6
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docstring-to-markdown==0.15
flake8==7.0.0
frozenlist==1.4.1
idna==3.7
importlib_metadata==7.2.1
isort==5.13.2
jedi==0.19.1
kombu==5.3.7
mccabe==0.7.0
multidict==6.0.5
packaging==24.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
pycodestyle==2.11.1
pydocstyle==6.3.0
pyflakes==3.2.0
PyJWT==2.8.0
pylint==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
pytoolconfig==1.3.1
redis==5.0.6
rope==1.13.0
six==1.16.0
slackclient==2.9.4
snowballstemmer==2.2.0
sqlparse==0.5.0
tomli==2.0.1
tomlkit==0.12.5
tzdata==2024.1
ujson==5.10.0
vine==5.1.0
wcwidth==0.2.13
whatthepatch==1.0.5
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
